fix(anim): fix lv_anim_set_repeat_count configuration parameter may be truncated

+++ src/widgets/animimage/lv_animimage.h
@@ -93,7 +93,7 @@ void lv_animimg_set_duration(lv_obj_t * img, uint32_t duration);
  * @param img       pointer to an animation image object
  * @param count     the number of times to repeat the animation
  */
-void lv_animimg_set_repeat_count(lv_obj_t * img, uint32_t count);
+void lv_animimg_set_repeat_count(lv_obj_t * img, uint16_t count);
 
 /*=====================
  * Getter functions
@@ -125,7 +125,7 @@ uint32_t lv_animimg_get_duration(lv_obj_t * img);
  * @param img   pointer to an animation image object
  * @return      the repeat count
  */
-uint32_t lv_animimg_get_repeat_count(lv_obj_t * img);
+uint16_t lv_animimg_get_repeat_count(lv_obj_t * img);
 
 #endif /*LV_USE_ANIMIMG*/
